requirements.txt
# Core Django
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1

# Database
psycopg2-binary==2.9.11

# Filtering & Search
django-filter==25.2

# API Documentation
drf-yasg==1.21.11

# Testing
pytest==9.0.2
pytest-django==4.11.1
pytest-cov==7.0.0
factory-boy==3.3.3

# Code Quality
black==25.12.0
isort==7.0.0
flake8==7.3.0
mypy==1.19.0
django-stubs==5.2.8
pre-commit==4.5.0

# Production
gunicorn==23.0.0
whitenoise==6.11.0

# Environment
python-dotenv==1.2.1

